+++ lib/src/cli/agent_cli.dart
@@ -16,8 +16,6 @@ library;
 
 import 'dart:async';
 
-import 'ansi_markdown.dart';
-
 import 'package:http/http.dart' as http;
 import 'package:meta/meta.dart';
 import 'package:yaml/yaml.dart';